Унаследовал вьюшки от абстрактного класса
Вынос метода renderPoint в файл utils.js оказался преждевременным.
При попытке запустить проект, страницы не отображались. Возможно, появились
циклические ссылки, но это не точно.
