+++ src/utils.js
@@ -1,10 +1,7 @@
-import Point from './view/point'
-import EditPoint from './view/edit-point';
-
 export const RenderPosition = {
   AFTERBEGIN: `afterbegin`,
   BEFOREEND: `beforeend`
-};
+}
 
 export const render = (container, element, place) => {
   switch (place) {
@@ -15,43 +12,10 @@ export const render = (container, element, place) => {
       container.append(element);
       break;
   }
-};
+}
 
 export const renderTemplate = (container, template, place) => {
   container.insertAdjacentHTML(place, template);
-};
-
-export const renderPoint = (container, point) => {
-  const pointComponent = new Point(point);
-  const pointEditComponent = new EditPoint(point);
-
-  const replacePointToForm = () => {
-    container.replaceChild(pointEditComponent.getElement(), pointComponent.getElement());
-  }
-
-  const replaceFormToPoint = () => {
-    container.replaceChild(pointComponent.getElement(), pointEditComponent.getElement());
-  }
-
-  pointComponent.getElement()
-    .querySelector(`.event__rollup-btn`)
-    .addEventListener(`click`, () => {
-      replacePointToForm();
-  });
-
-  pointEditComponent.getElement()
-    .querySelector(`form`)
-    .addEventListener(`submit`, () => {
-      replaceFormToPoint();
-    });
-
-  pointEditComponent.getElement()
-    .querySelector(`.event__rollup-btn`)
-    .addEventListener(`click`, () => {
-      replaceFormToPoint();
-    });
-
-  render(container, pointComponent.getElement(), RenderPosition.BEFOREEND);
 }
 
 export const createElement = (template) => {
@@ -59,4 +23,4 @@ export const createElement = (template) => {
   newElement.innerHTML = template;
 
   return newElement.firstChild;
-};
+}
