+++ src/view/events-sort-form.js
@@ -1,8 +1,7 @@
-import {createElement} from '../utils';
+import AbstractView from './abstract';
 
 function createTripEventsSortForm() {
-  return `
-    <form class="trip-events__trip-sort  trip-sort" action="#" method="get">
+  return `<form class="trip-events__trip-sort  trip-sort" action="#" method="get">
       <div class="trip-sort__item  trip-sort__item--day">
         <input id="sort-day" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort" value="sort-day" checked>
         <label class="trip-sort__btn" for="sort-day">Day</label>
@@ -30,24 +29,8 @@ function createTripEventsSortForm() {
     </form>`;
 }
 
-export default class EventsSortForm {
-  constructor() {
-    this._element = null;
-  }
-
+export default class EventsSortForm extends AbstractView{
   getTemplate() {
     return createTripEventsSortForm();
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
   }
 }
