+++ src/view/events.js
@@ -1,4 +1,4 @@
-import {createElement} from '../utils';
+import AbstractView from './abstract';
 
 function createTripEventsList() {
   return `<ul class="trip-events__list">
@@ -6,24 +6,8 @@ function createTripEventsList() {
 }
 
 
-export default class Events {
-  constructor() {
-    this._element = null;
-  }
-
+export default class Events extends AbstractView{
   getTemplate() {
     return createTripEventsList();
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
   }
 }
