+++ src/view/info-date.js
@@ -1,28 +1,17 @@
-import {createElement} from '../utils';
+import AbstractView from './abstract';
 
 function createTripInfoDate(dateRange) {
     return `<p class="trip-info__dates">${dateRange}</p>`;
 }
 
-export default class InfoDate {
+export default class InfoDate extends AbstractView{
   constructor(dateRange) {
+    super();
+
     this._dateRange = dateRange;
-    this._element = null;
   }
 
   getTemplate() {
     return createTripInfoDate(this._dateRange);
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
   }
 }
