+++ src/view/switches.js
@@ -1,31 +1,14 @@
-import {createElement} from '../utils';
+import AbstractView from './abstract';
 
 function createTripSwitches() {
-  return `
-    <nav class="trip-controls__trip-tabs  trip-tabs">
+  return `<nav class="trip-controls__trip-tabs  trip-tabs">
       <a class="trip-tabs__btn  trip-tabs__btn--active" href="#">Table</a>
       <a class="trip-tabs__btn" href="#">Stats</a>
     </nav>`;
 }
 
-export default class Switches {
-  constructor() {
-    this._element = null;
-  }
-
+export default class Switches extends AbstractView {
   getTemplate() {
     return createTripSwitches();
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
   }
 }
